app/requirements.txt
requests==2.31.0
flask==2.3.2
openai==0.27.8
pyyaml==6.0.1
